pyproject.toml
[project]
name = "agentic-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "pdf2img>=0.1.2",
    "surya-ocr>=0.17.0",
    "tqdm>=4.67.1",
]
